some refactor cleanup prepping for creating indirect mode controller and support Moved witness stuff to indirecting
