+++ tests/core/test_replay.py
@@ -11,7 +11,7 @@ import pytest
 from keri import help
 from keri.help import helping
 from keri.db import dbing
-from keri.base import keeping, directing
+from keri.base import basing, keeping, directing
 from keri.core import coring, eventing
 
 
@@ -36,7 +36,7 @@ def test_replay():
 
         # setup Deb's habitat using default salt multisig already incepts
         sith = ["1/2", "1/2", "1/2"]  # weighted signing threshold
-        debHab = directing.Habitat(ks=debKS, db=debDB, isith=sith, icount=3,
+        debHab = basing.Habitat(ks=debKS, db=debDB, isith=sith, icount=3,
                                    temp=True)
         assert debHab.ks == debKS
         assert debHab.db == debDB
@@ -45,7 +45,7 @@ def test_replay():
         # setup Cam's habitat using default salt multisig already incepts
         # Cam's receipts will be vrcs with 3 indexed sigantures attached
         sith = '2'  # hex str of threshold int
-        camHab = directing.Habitat(ks=camKS, db=camDB, isith=sith, icount=3,
+        camHab = basing.Habitat(ks=camKS, db=camDB, isith=sith, icount=3,
                                    temp=True)
         assert camHab.ks == camKS
         assert camHab.db == camDB
@@ -54,7 +54,7 @@ def test_replay():
         # setup Bev's habitat using default salt nonstransferable already incepts
         # Bev's receipts will be rcts with a receipt couple attached
         sith = '1'  # hex str of threshold int
-        bevHab = directing.Habitat(ks=bevKS, db=bevDB, isith=sith, icount=1,
+        bevHab = basing.Habitat(ks=bevKS, db=bevDB, isith=sith, icount=1,
                                    transferable=False, temp=True)
         assert bevHab.ks == bevKS
         assert bevHab.db == bevDB
@@ -65,7 +65,7 @@ def test_replay():
         # Art's receipts will be rcts with a receipt couple attached
         salt = coring.Salter(raw=b'abcdef0123456789').qb64
         sith = '1'  # hex str of threshold int
-        artHab = directing.Habitat(ks=artKS, db=artDB, isith=sith, icount=1,
+        artHab = basing.Habitat(ks=artKS, db=artDB, isith=sith, icount=1,
                                    salt=salt, transferable=False, temp=True)
         assert artHab.ks == artKS
         assert artHab.db == artDB
@@ -510,7 +510,7 @@ def test_replay_all():
 
         # setup Deb's habitat using default salt multisig already incepts
         sith = ["1/2", "1/2", "1/2"]  # weighted signing threshold
-        debHab = directing.Habitat(ks=debKS, db=debDB, isith=sith, icount=3,
+        debHab = basing.Habitat(ks=debKS, db=debDB, isith=sith, icount=3,
                                    temp=True)
         assert debHab.ks == debKS
         assert debHab.db == debDB
@@ -519,7 +519,7 @@ def test_replay_all():
         # setup Cam's habitat using default salt multisig already incepts
         # Cam's receipts will be vrcs with 3 indexed sigantures attached
         sith = '2'  # hex str of threshold int
-        camHab = directing.Habitat(ks=camKS, db=camDB, isith=sith, icount=3,
+        camHab = basing.Habitat(ks=camKS, db=camDB, isith=sith, icount=3,
                                    temp=True)
         assert camHab.ks == camKS
         assert camHab.db == camDB
@@ -528,7 +528,7 @@ def test_replay_all():
         # setup Bev's habitat using default salt nonstransferable already incepts
         # Bev's receipts will be rcts with a receipt couple attached
         sith = '1'  # hex str of threshold int
-        bevHab = directing.Habitat(ks=bevKS, db=bevDB, isith=sith, icount=1,
+        bevHab = basing.Habitat(ks=bevKS, db=bevDB, isith=sith, icount=1,
                                    transferable=False, temp=True)
         assert bevHab.ks == bevKS
         assert bevHab.db == bevDB
@@ -539,7 +539,7 @@ def test_replay_all():
         # Art's receipts will be rcts with a receipt couple attached
         salt = coring.Salter(raw=b'abcdef0123456789').qb64
         sith = '1'  # hex str of threshold int
-        artHab = directing.Habitat(ks=artKS, db=artDB, isith=sith, icount=1,
+        artHab = basing.Habitat(ks=artKS, db=artDB, isith=sith, icount=1,
                                    salt=salt, transferable=False, temp=True)
         assert artHab.ks == artKS
         assert artHab.db == artDB
