+++ tests/demo/test_demo.py
@@ -9,7 +9,7 @@ import logging
 from hio.base import doing
 from hio.core.tcp import clienting, serving
 
-from keri.base import directing, keeping
+from keri.base import basing, keeping, directing
 from keri.db import dbing
 from keri.core import eventing, coring
 from keri.demo import demoing
@@ -92,7 +92,7 @@ def test_direct_mode_bob_eve_demo():
         eveKevers = dict()
 
         # setup bob
-        bobHab = directing.Habitat(name='Bob', ks=bobKS, db=bobDB, kevers=bobKevers,
+        bobHab = basing.Habitat(name='Bob', ks=bobKS, db=bobDB, kevers=bobKevers,
                                    secrecies=bobSecrecies, temp=True)
         assert bobHab.ks == bobKS
         assert bobHab.db == bobDB
@@ -125,7 +125,7 @@ def test_direct_mode_bob_eve_demo():
         # Bob's Reactants created on demand
 
         # setup eve
-        eveHab = directing.Habitat(name='Eve', ks=eveKS,  db=eveDB, kevers=eveKevers,
+        eveHab = basing.Habitat(name='Eve', ks=eveKS,  db=eveDB, kevers=eveKevers,
                                    secrecies=eveSecrecies, temp=True)
 
         assert eveHab.ks == eveKS
@@ -260,7 +260,7 @@ def test_direct_mode_sam_eve_demo():
         eveKevers = dict()
 
         # setup Sam
-        samHab = directing.Habitat(name='Sam', ks=samKS,  db=samDB, kevers=samKevers,
+        samHab = basing.Habitat(name='Sam', ks=samKS,  db=samDB, kevers=samKevers,
                                    secrecies=samSecrecies, temp=True)
 
         assert samHab.ks == samKS
@@ -294,7 +294,7 @@ def test_direct_mode_sam_eve_demo():
         # Sam's Reactants created on demand
 
         # setup eve
-        eveHab = directing.Habitat(name='Eve', ks=eveKS,  db=eveDB, kevers=eveKevers,
+        eveHab = basing.Habitat(name='Eve', ks=eveKS,  db=eveDB, kevers=eveKevers,
                                    secrecies=eveSecrecies, temp=True)
         assert eveHab.ks == eveKS
         assert eveHab.db == eveDB
@@ -537,7 +537,7 @@ def test_indirect_mode_sam_cam_wit_demo():
         witKevers = dict()
 
         # setup the witness
-        witHab = directing.Habitat(name='Wit', ks=witKS, db=witDB, kevers=witKevers, isith=1, icount=1,
+        witHab = basing.Habitat(name='Wit', ks=witKS, db=witDB, kevers=witKevers, isith=1, icount=1,
                                    temp=True, transferable=False)
         wit = witHab.pre
         assert witHab.ks == witKS
@@ -567,7 +567,7 @@ def test_indirect_mode_sam_cam_wit_demo():
         assert sam == 'EhnaYUqhHoo8kZaXJuTDCZ-h5ZDx3st4NemgqSITWp48'
 
 
-        samHab = directing.Habitat(name='Sam', ks=samKS, db=samDB, kevers=samKevers, wits=[wit],
+        samHab = basing.Habitat(name='Sam', ks=samKS, db=samDB, kevers=samKevers, wits=[wit],
                                    secrecies=samSecrecies, temp=True)
         assert samHab.ks == samKS
         assert samHab.db == samDB
@@ -600,7 +600,7 @@ def test_indirect_mode_sam_cam_wit_demo():
         # Sam's Reactants created on demand
 
         # setup cam
-        camHab = directing.Habitat(name='Cam', ks=camKS,  db=camDB, kevers=camKevers,
+        camHab = basing.Habitat(name='Cam', ks=camKS,  db=camDB, kevers=camKevers,
                                    secrecies=camSecrecies, temp=True)
 
         assert camHab.ks == camKS
